@@ -17,14 +17,13 @@ class AppServiceProvider extends ServiceProvider
         //
     }
 
-
     /**
      * Bootstrap any application services.
      *
      * @return void
      */
     public function boot()
-     {
-    Schema::defaultStringLength(191);
+    {
+        Schema::defaultStringLength(191);
     }
 }
