@@ -1,111 +1,248 @@
-@extends('layouts.app_dashbord') @section('content')
-<link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel="stylesheet" />
-<link rel="stylesheet" type="text/css" href={{ asset('assets_dashbord/css_vue/materialdesignicons.min.css') }}>
-<link rel="stylesheet" type="text/css" href={{ asset('assets_dashbord/css_vue/vuetify.min.css') }}>
-<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=no, minimal-ui" />
-<div class="d-sm-flex align-items-center justify-content-between mb-4">
-   <h1 class="h3 mb-0 text-gray-800">Gérer les gategorie</h1>
-</div>
-<div class="row">
-   <div class="col-md-12">
-      <div id="app_gategorie">
-         <template>
-            <v-row>
-               <v-col cols="4" sm="5">
-                  <v-card>
-                     <v-toolbar
-                        color="pink"
-                        dark
-                        dense
-                        flat
-                        >
-                        <v-toolbar-title class="body-2">Ajouter un gategorie</v-toolbar-title>
-                     </v-toolbar>
-                     <v-card-text>
-                        <v-form @submit.prevent="add">
-                           <v-row>
-                              <v-container>
-                                 <v-text-field label="Nom" single-line solo v-model="gategorie_a.nom" required></v-text-field>
-                                 <v-text-field label="" single-line solo-inverted v-model="gategorie_a.date_create" disabled></v-text-field>
-                                 <v-btn color="success" class="mr-4" type="submit">
-                                    Ajouter
-                                 </v-btn>
-                                 <v-btn color="error" class="mr-4" @click="reset">
-                                    Effacer
-                                 </v-btn>
-                              </v-container>
-                           </v-row>
-                        </v-form>
-                     </v-card-text>
-                  </v-card>
-               </v-col>
-               <v-col cols="8" sm="7">
-                  <v-card>
-                     <v-card-title>
-                        La liste des gategories
-                        <v-spacer></v-spacer>
-                        <v-text-field  v-model="search" append-icon="mdi-magnify" label="Search" single-line hide-details></v-text-field>
-                     </v-card-title>
-                     <v-col cols="12" class="pl-remove" sm="3" pb="2"  v-if="btn_control">
-                        <v-btn class="btn_remove" depressed @click="deleteItem()">
-                           <v-icon class="color-icon-remove  ">mdi-delete</v-icon>
-                        </v-btn>
-                     </v-col>
-                     <v-data-table  @input="item($event)" :headers="headers" :items="gategorie" :search="search" :value="selectedRows" v-model="selected" :items-per-page="5"  :sort-by.sync="sortBy"
-                        :sort-desc.sync="sortDesc" show-select   item-key="id"
-                        :expanded.sync="expanded" @click:row="clicked">
-                        <template v-slot:item.action="{ item }">
-                           <v-btn color="purple" fab small dark  @click="editItem(item)">
-                              <v-icon>mdi-pencil</v-icon>
-                           </v-btn>
-                           <v-dialog v-model="dialog" max-width="500px">
-                              <v-card>
-                                 <v-card-title>
-                                    <span class="headline">Modifier la gategorie</span>
-                                 </v-card-title>
-                                 <v-container>
-                                    <v-row class="pl-3 pr-3" >
-                                       <v-col cols="12" sm="6" md="12">
-                                          <v-text-field pl="5" v-model="editedItem.nom"  label="Nom"></v-text-field>
-                                       </v-col>
-                                       <v-col cols="12" sm="6" md="12">
-                                          <v-text-field  v-model="editedItem.date_create" label="Date" disabled></v-text-field>
-                                       </v-col>
-                                    </v-row>
-                                 </v-container>
-                                 </v-form>
-                                 </v-card-title>
-                                 <v-card-actions>
-                                    <v-spacer></v-spacer>
-                                    <v-btn color="blue darken-1" text @click="close">Annuler</v-btn>
-                                    <v-btn color="blue darken-1" text @click="save">Sauvegarder
-                                    </v-btn>
-                                 </v-card-actions>
-                              </v-card>
-                           </v-dialog>
-                        </template>
-         </template>
-         </v-data-table>
-         </v-card>
-         </v-col>
-         </v-row>
-         </template>
-      </div>
-   </div>
-</div>
-<script src="{{ asset('js/plugins/vue.js') }}"></script>
-<script src="{{ asset('js/plugins/vee-validate.js') }}"></script>
-<script src="{{ asset('js/plugins/axios.min.js') }}"></script>
-<script src="{{ asset('js/plugins/sweetalert2@9.js') }}"></script>
-<script src="{{ asset('js/plugins/vuetify.js') }}"></script>
-<script>
-   window.laravel ={!! json_encode([
-     'token' => csrf_token(),
-     'url'   => url('/'),
-     'date'   => date('Y-m-d'),
+@extends('layouts.master')
+@section('main-content')
+
+           <link rel="stylesheet" type="text/css" href={{ asset('css_vuetify/materialdesignicons.min.css') }}>
+           <link rel="stylesheet" type="text/css" href={{ asset('css_vuetify/vuetify.min.css') }}>
+           <link rel="stylesheet" type="text/css" href={{ asset('assets/styles/css/custom_vuetify.css') }}>
+           
+           <div class="breadcrumb">
+                <h1>  La liste des Gategories</h1>
+            
+            </div>
+
+            <div class=" border-top"></div>
+
+        
+
+            <div id="app_gategorie" data-app>
+         
+
+               <div class="row">
+                  <div class="col-md-12">
+                      <div class="card">
+                          <div class="card-header  gradient-purple-indigo  0-hidden pb-80">
+                             <div class="pt-4">
+                              <div class="row">
+                                  <h4 class="col-md-4 text-white">Gategories</h4>
+                                  <input v-model="search" type="text" class="form-control form-control-rounded col-md-4 ml-3 mr-3"  append-icon="mdi-magnify" placeholder="Rechercher Gategories ...">
+                                  <i aria-hidden="true" class="v-icon notranslate btn_search mdi mdi-magnify theme--light"></i>
+                              </div>
+                             </div>
+                          </div>
+                          <div class="card-body">
+                              <div class="ul-contact-list-body">
+                                  <div class="ul-contact-main-content">
+                                      <div class="ul-contact-left-side">
+                                          <div class="card">
+                                              <div class="card-body">
+                                                  <div class="ul-contact-list">
+                                                      <div class="contact-close-mobile-icon float-right mb-2">
+                                                          <i class="i-Close-Window text-15 font-weight-600"></i>
+                                                      </div>
+                                                      <!-- modal  -->
+                                                      
+                                                      <button class="btn btn-outline-secondary btn-block mb-4" >
+                                                        Parametres
+                                                       </button>
+
+                                                      <template>
+                                                         <v-row justify="center">
+                                                           <v-dialog
+                                                             v-model="dialog_add"
+                                                             persistent
+                                                             max-width="600px"
+                                                           >
+                                                            
+                                                             <v-card>
+                                                               <v-form @submit.prevent="add">
+                                                               <v-card-title>
+                                                                 <span class="headline">Ajouter un gategorie</span>
+                                                               </v-card-title>
+                                                               <v-card-text>
+                                                                 <v-container>
+                                                                
+                                                                   <v-row>
+                                                                     <v-col cols="12" sm="6" md="12"  >
+
+                                                                        <v-text-field
+                                                                         label="Nom*"
+                                                                         v-model="gategorie_a.nom"
+                                                                         required
+                                                                       ></v-text-field>
+
+                                                                     </v-col>
+                                                
+                                                                      
+                                                                    
+                                                                    
+                                                                     <v-col cols="12" sm="6" md="12" >
+
+                                                                        <v-text-field
+                                                                        label="date creation*"
+                                                                        v-model="gategorie_a.date_create"
+                                                                        required
+                                                                        disabled
+                                                                      ></v-text-field>
+                                                                      
+                                                                    </v-col>
+                                                                      
+                                                                     
+                                                                    
+                                                                  
+                                                                  
+                                                                    
+                                                                     
+                                                                   </v-row>
+                                                                 
+                                                                 </v-container>
+                                                                 <small>* indique le champ obligatoire</small>
+                                                               </v-card-text>
+                                                               <v-card-actions>
+                                                                 <v-spacer></v-spacer>
+                                                                 <v-btn
+                                                                 color="primary mr-3"
+                                                                 dark
+                                                                  
+                                                                   @click="dialog_add = false"
+                                                                 >
+                                                                   Fermer 
+                                                                 </v-btn>
+                                                                 <v-btn color="error" class="mr-4" @click="reset">
+                                                                  Effacer
+                                                                  </v-btn>
+                                                                 <v-btn color="success"   class="mr-4" type="submit">
+                                                                  Ajouter
+                                                                 </v-btn>
+                                                               </v-card-actions>
+                                                              </v-form>
+                                                             </v-card>
+                                                           </v-dialog>
+                                                         </v-row>
+                                                       </template>   
+                                                      
+   
+                                                      <!-- end:modal  -->
    
    
-   ]) !!}
-</script>
-<script src="{{ asset('js/gategories_vue.js') }}"></script>
+                                                      <div class="list-group" id="list-tab" role="tablist">
+                                                          <a    @click="dialog_add = true" class="list-group-item list-group-item-action border-0" id="list-home-list" data-toggle="list" href="#list-home" role="tab" aria-controls="home" aria-selected="false">
+                                                              <i class="nav-icon i-Add-Window"></i>
+                                                              Nouveau Gategorie</a>
+                                                       
+                                                          <a  @click="remove_item" class="list-group-item list-group-item-action border-0" id="list-settings-list" data-toggle="list" href="#list-settings" role="tab" aria-controls="settings" aria-selected="false">
+                                                              <i class="nav-icon i-Remove"></i>
+                                                              Supprimer</a>
+   
+                                                          <label for="" class="text-muted font-weight-600 py-8">MEMBERS</label>
+   
+                                                          <a class="list-group-item list-group-item-action border-0 " id="list-home-list" data-toggle="list" href="#list-home" role="tab" aria-controls="home">
+                                                              <i class="nav-icon i-Arrow-Next"></i>
+                                                              Contact List</a>
+                                                          <a class="list-group-item list-group-item-action border-0 active show" id="list-profile-list" data-toggle="list" href="#list-profile" role="tab" aria-controls="profile" aria-selected="true">
+                                                              <i class="nav-icon i-Arrow-Next"></i>
+                                                              Conected</a>
+                                                          <a class="list-group-item list-group-item-action border-0" id="list-settings-list" data-toggle="list" href="#list-settings" role="tab" aria-controls="settings" aria-selected="false">
+                                                              <i class="nav-icon i-Arrow-Next"></i>
+                                                              Settings</a>
+                                                      </div>
+                                                  </div>
+                                              </div>
+                                          </div>
+                                      </div>
+                                      <div class="ul-contact-content">
+                                          <div class="card">
+                                            
+                                             <v-card>
+                                                <v-card-title>
+                                                 la liste des Gategories
+                                                   <v-spacer></v-spacer>
+                                                   
+                                                </v-card-title>
+                                              
+                                                <v-data-table  @input="item($event)" :headers="headers" :items="gategorie" :search="search" :value="selectedRows" v-model="selected" :items-per-page="5"  :sort-by.sync="sortBy"
+                                                :sort-desc.sync="sortDesc" show-select   item-key="id"
+                                                :expanded.sync="expanded" @click:row="clicked">
+                                                <template v-slot:item.action="{ item }">
+                                                   <v-btn color="purple" fab small dark  @click="editItem(item)">
+                                                      <i class="nav-icon i-Pen-2 font-weight-bold"></i>
+                                                   </v-btn>
+                                                   <v-dialog v-model="dialog" max-width="500px">
+                                                      <v-card>
+                                                         <v-card-title>
+                                                            <span class="headline">Modifier la gategorie</span>
+                                                         </v-card-title>
+                                                         <v-container>
+                                                            <v-row class="pl-3 pr-3" >
+                                                               <v-col cols="12" sm="6" md="12">
+                                                                  <v-text-field pl="5" v-model="editedItem.nom"  label="Nom"></v-text-field>
+                                                               </v-col>
+                                                               <v-col cols="12" sm="6" md="12">
+                                                                  <v-text-field  v-model="editedItem.date_create" label="Date" disabled></v-text-field>
+                                                               </v-col>
+                                                            </v-row>
+                                                         </v-container>
+                                                         </v-form>
+                                                         </v-card-title>
+                                                         <v-card-actions>
+                                                            <v-spacer></v-spacer>
+                                                            <v-btn color="blue darken-1" text @click="close">Annuler</v-btn>
+                                                            <v-btn color="blue darken-1" text @click="save">Sauvegarder
+                                                            </v-btn>
+                                                         </v-card-actions>
+                                                      </v-card>
+                                                   </v-dialog>
+                                                </template>
+                                                
+                                                </v-data-table>
+                                              
+                                             </v-card>
+                                             
+                                          </div>
+                                      </div>
+                                  </div>
+                              </div>
+                          </div>
+                      </div>
+                  </div>
+              </div>
+            </div>
+
+
+          
+
+            
+
+            
+
+
 @endsection
+
+@section('page-js')
+
+            <script src="{{ asset('js/plugins/vue.js') }}"></script>
+            <script src="{{ asset('js/plugins/vee-validate.js') }}"></script>
+            <script src="{{ asset('js/plugins/axios.min.js') }}"></script>
+            <script src="{{ asset('js/plugins/sweetalert2@9.js') }}"></script>
+            <script src="{{ asset('js/plugins/vuetify.js') }}"></script>
+            <script>
+               window.laravel ={!! json_encode([
+                 'token' => csrf_token(),
+                 'url'   => url('/'),
+                 'date'   => date('Y-m-d'),
+               
+               
+               ]) !!}
+            </script>
+          <script src="{{ asset('js/gategories_vue.js') }}"></script>
+
+@endsection
+
+
+
+
+
+
+
+   
+
+
