@@ -1,159 +1,237 @@
-@extends('layouts.app_dashbord') @section('content')
-<link href="https://fonts.googleapis.com/css?family=Roboto:100,300,400,500,700,900" rel="stylesheet" />
+@extends('layouts.master')
+@section('page-css')
+
+<style>
+  .table-product.active{
+   display: none;
+  }
+  td.row_p {
+    width: 214px;
+  }
+  
+</style>
+
+@endsection
+
+@section('main-content')
 
 
-
-<div class="d-sm-flex align-items-center justify-content-between mb-4">
-    <script src={{ asset('assets/vendor/jquery/jquery-3.2.1.min.js') }}></script>
-   <script src="{{ asset('js/plugins/select2.min.js') }}"></script>
-   <h1 class="h3 mb-0 text-gray-800">Commande</h1>
+<div class="breadcrumb">
+    <h1>Checkout</h1>
+    <ul>
+        <li><a href="">Apps</a></li>
+        <li>Ecommerce</li>
+    </ul>
 </div>
-<div class="row">
-   <div class="col-md-12">
-      <!-- Panel order -->
-      <div class="card shadow mb-4 p-0" >
-         <div class="card-header py-3">
-            <div class="profile-description"></div>
-            <h6 class="m-0 titre-panel">Ajouter un commande</h6>
-         </div>
-      
-         <div class="card-body">
-           <form action="{{ url('order')}}" method="POST" id="FormOrder" onsubmit="return validateForm();">
-               {{ csrf_field() }}
-            <div class="row">
-               <div class="col-md-12">
+<div class="separator-breadcrumb border-top"></div>
+
+<form action="{{ url('order')}}" method="POST" id="FormOrder" onsubmit="return validateForm();">
+  {{ csrf_field() }}
+<section class="chekout-page">
+    <div class="row">
+      <div class="col-lg-6 mb-6">
+        <div class="card">
+          <div class="card-body">
+            <div class="d-flex justify-content-between">
+              <div class="card-title">Panier de produit</div>
+              <div class="headder-elements tt">
+                <div class="list-icons">
+                    <a href="" class="ul-task-manager__list-icon " id="arrow-down"><i class="i-Arrow-Down"></i></a>
+                    <a href="" class="ul-task-manager__list-icon btn-add"><i class="i-Add"></i></a>
+
+                </div>
+            </div>
              
-                  <label for="titre" class="label-p">Date de commande  </label>
-                  <input type="date" name="date_commande" class="form-control input-product">
-               </div>
-               <div class="col-md-12 pt-2">
-                  <div class="half-form pr-3">
-                     <label for="quantite" class="label-p">Nom de client </label>
-                     <select id="valueclient" onchange="getphone()" class="select-js-client form-control input-product" style="width: 100%" name="idclient" required>
-                      <option class="select2-results__group"  value="">Selectionner le Client</option>
-                      @foreach ($clients as $client)
-                      <option  value="{{$client->id}}">{{$client->nom}}</option>
-                      @endforeach
-                   </select>
-                  </div>
-                  <div class="half-form pl-3">
-                     <label  class="label-p">Numero Telephone de client </label>
-                     <input type="text" id="phone_client"  class="form-control input-product" disabled required>
-                  </div>
-               </div>
-               <div class="col-md-12 pt-2">
-                
-                  <div class="col-md-12 pt-2">
-                     <table class="table table-bordered" id="table">
-                        <thead>
-                           
-                           <tr >
-                              <th style="width:30%" scope="col">Produit</th>
-                              <th scope="col">Prix</th>
-                              <th scope="col">Quantite</th>
-                              <th scope="col">Total</th>
-                              <th>
-                              </th>
-                           </tr>
-                        </thead>
-                        <tr id='row1'>
-                           <td>
-                              <select  id="productName1" onchange="getProductData(1)" class="js-example-basic-single form-control input-product" style="width: 100%" name="product[]" required>
-                                 <option class="select2-results__group"  value="">Selectionner le produit</option>
-                                 @foreach ($products as $product)
-                                 <option  value="{{$product->id}}">{{$product->titre}}</option>
-                                 @endforeach
-                              </select>
-                           </td>
-                           <td id="col2"><input type="number"  id="rate1" class="form-control input_or"  disabled required >
-                            <input type="hidden" id="rateValue1" name="rate[]" class="form-control input_or" required>
-                           </td>
-                           <td id="col3"><input type="number" onclick="calcul_total(1)" id="quantite1" name="quantite[]" class="form-control input_or" required>
-                           
-                           </td>
-                           <td id="col4"><input type="number" id="total1" name="total[]" class="form-control input_or" disabled required>
-                            <input type="hidden" id="totalValue1" name="totalp[]" class="form-control input_or" required>
-                           </td>
-                           <td >
-                              <a href="#" class="btn btn-danger btn-circle btn-remove" onClick="removeRow(1)" required>
-                              <i class="fa fa-trash"></i>
-                              </a>
-                           </td>
-                        </tr>
-                
-                     </table>
-                     <div class="d-block text-right">
-                      <a href="#" class="btn-add btn-add" >
-                      <i class="zmdi zmdi-plus-circle btn-success1"></i>
-                      <span class="text-add-row">ajouter nouveau ligne</span>
-                      </a>
-                   </div>
-                  </div>
-                  <div class="col-md-12 pt-2">
-                     <div class="half-form pr-3">
-                        <label for="quantite" class="label-p">Sous Total </label>
-                        <input type="text" id="subTotal" name="subTotal" class="form-control input-product" disabled required>
-                        <input type="hidden" id="subTotalvalue" name="subTotalvalue" class="form-control input-product" required>
-                        <label for="tva" class="label-p">Tva </label>
-                        <input type="text" id="tva" name="tva" class="form-control input-product" disabled required>
-                        <input type="hidden" id="tvavalue" name="tvavalue"  >
-                        <label for="total" class="label-p">Total </label>
-                        <input type="text" id="total" name="total" class="form-control input-product" disabled required>
-                        <input type="hidden" id="totalvalue" name="total" >
-                     </div>
-                     <div class="half-form pl-3">
-                        <label for="typepaiement" class="label-p">Type de paiement </label>
-                        <select  class="form-control input-product" style="width: 100%" name="typepaiement" required>
-                           <option  value="">Selectionner </option>
-                           <option value="Cheque">Cheque</option>
-                           <option value="Cash">Cash</option>
-                           <option value="Credit Card">Credit Card</option>
-                           <option  value="Cash">Cash</option>
-                        </select>
-                        <label for="statutpaiement" class="label-p ">Statut de paiement </label>
-                        <select  class="form-control input-product" style="width: 100%" name="statutpaiement" required>
-                           <option  value="">Selectionner</option>
-                           <option value="Règlement de la totalité">Règlement de la totalité</option>
-                           <option value="paiement anticipét">paiement anticipét</option>
-                           <option value="aucun paiement">aucun paiement</option>
-                        </select>
-                     </div>
-                     
-                  </div>
-                  
-               </div>
             </div>
-        
-            <div class="col-md-12 pt-3">
-               <div class="text-center">
-                <button type="submit" class="btn btn-success btn-icon-split">
-                  <span class="icon text-white-50">
-                    <i class="fa fa-check"></i>
-                  </span>
-                  <span class="text mr-2">Validé</span>
+           
+
+
+            <div class="table-responsive table-product ">
+              <table id="table_product" class="table">
+                <thead>
+                  <tr>
+                    <th scope="col">Product</th>
+                    <th scope="col">Price</th>
+                    <th scope="col">Quantity</th>
+                    <th scope="col">Total</th>
+                    <th scope="col">Action</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  <tr id='row1'>
+                    <td scope="row" class="row_p">
+                      <select  id="productName1" onchange="getProductData(1)" class="js-example-basic-single form-control input-product" style="width: 100%" name="product[]" required>
+                        <option   value="">Selectionner le produit</option>
+                        @foreach ($products as $product)
+                        <option  value="{{$product->id}}">{{$product->titre}}</option>
+                        @endforeach
+                     </select>
+                    </td>
+                    <td> <div class="d-flex "> <div id="price1"> 0  </div>  <span class="text-muted pl-1">DH</span>  </div> 
+                      <input type="hidden" id="priceValue1" name="rate[]" class="form-control input_or" >
+                    </td>
+                    <td>
+                       <input type="number" onclick="calcul_total(1)" value="0" id="quantite1" name="quantite[]" class="form-control input_or" required>
+                      
+                    </td>
+                    <td><div class="d-flex "> <div id="total_product1"> 0.00  </div>  <span class="text-muted pl-1">DH</span>
+                      <input type="hidden" id="total_productValue1" name="totalp[]" class="form-control input_or">
+                     </td>
+                    <td>
+                      <a href="" onClick="removeRow(event,1)" ><i class="i-Close-Window text-19 text-danger font-weight-700 prevent-default"></i></a>
+                    
+                    </td>
+                  </tr>
+                </tbody>
+              </table>
+            </div>
+
+            <div class="row ">
+              <div class="col-lg-12 mt-5">
+                <div class="ul-product-cart__invoice">
+                  <div class="card-title">
+                    <h4 class="heading text-primary">Total Payment</h4>
+                  </div>
+                  <table class="table">
+                    <tbody>
+                      <tr>
+                        <th scope="row" class="text-16">Sous Total</th>
+                        <td class="text-16 text-success font-weight-700">
+                          <div class="d-flex "> <div id="subTotal"> 0.00  </div>  <span class=" pl-1">DH</span>  </div>
+                          <input type="hidden" id="subTotalValue" name="subTotalvalue" class="form-control input-product" required>
+                        </td>
+                      </tr>
+                      <tr>
+                        <th scope="row" class="text-16">Tva</th>
+                        <td>
+                           <span>
+                           <div class="d-flex "> <div id="tva"> 0.00  </div>  <span class="text-muted pl-1">DH</span> 
+                           </div>
+                          </span>
+                          <input type="hidden" id="tvaValue" name="tvavalue">
+                        </td>
+                      </tr>
+                      <tr>
+                        <th scope="row" class="text-primary text-16">
+                          Total:
+                        </th>
+                        <td class="font-weight-700 text-16">
+                          <div class="d-flex "> <div id="total"> 0.00  </div>  <span class="text-muted pl-1">DH</span> 
+                          <input type="hidden" id="totalValue" name="total" >
+                        </td>
+                      </tr>
+                    </tbody>
+                  </table>
+                </div>
+                
+              </div>
+              
+            </div>
+          </div>
+          <div class="card-footer">
+            <div class="row text-right">
+              <div class="col-lg-12 ">
+                <button type="submit" class="btn btn-success m-1">
+                 Validé
                 </button>
-                <a href="#" id="" class="btn btn-secondary btn-icon-split btn-reset" onclick="resetOrderForm()">
-                  <span class="icon text-white-50">
-                    <i class="fa fa-arrow-right"></i>
-                  </span>
-                  <span class="text mr-2">Effacer</span>
-                </a>
-               </div>
+              </div>
             </div>
-          </form>
-         </div>
+          </div>
+        </div>
       </div>
-   </div>
-</div>
+      <div class="col-lg-6">
+        <div class="card">
+          <div class="card-body">
+         
+              <div class="card-body">
+                <div class="card-title">Les informations Supplémentaire de Commandes</div>
+
+                <div class="form-row">
+                  <div class="form-group col-md-12">
+                    <label for="inputtext11" class="ul-form__label">Date de commande:</label>
+                    <input type="date" name="date_commande" class="form-control" id="inputtext11" />
+                  </div>
+                </div>
+
+
+              
+
+                <div class="form-row">
+                  <div class="form-group col-md-6">
+                    <label for="inputtext14" class="ul-form__label">Nom de client:</label>
+                    <select id="valueclient" onchange="getphone()" class="select-js-client form-control input-product" style="width: 100%" name="idclient" required>
+                     <option class="select2-results__group"  value="">Selectionner le Client</option>
+                     @foreach ($clients as $client)
+                     <option  value="{{$client->id}}">{{$client->nom}}</option>
+                     @endforeach
+                  </select>
+                  </div>
+                  <div class="form-group col-md-6">
+                    <label for="inputEmail15" class="ul-form__label">Numero Telephone de client :</label>
+                    <input type="text" id="phone_client"  class="form-control" disabled required>
+                  </div>
+
+                 
+                </div>
+              </div>
+            
+          </div>
+        </div>
+
+        <div class="card mt-4">
+          <div class="card-body">
+            <div class="card-title">Mode de Paiment</div>
+            
+            <div class="half-form pl-3 pb-3">
+               <label for="typepaiement" class="label-p">Type de paiement </label>
+               <select  class="form-control input-product" style="width: 100%" name="typepaiement" required>
+                  <option  value="">Selectionner </option>
+                  <option value="Cheque">Cheque</option>
+                  <option value="Cash">Cash</option>
+                  <option value="Credit Card">Credit Card</option>
+                  <option  value="Cash">Cash</option>
+               </select>
+               <label for="statutpaiement" class="label-p ">Statut de paiement </label>
+               <select  class="form-control input-product" style="width: 100%" name="statutpaiement" required>
+                  <option  value="">Selectionner</option>
+                  <option value="Règlement de la totalité">Règlement de la totalité</option>
+                  <option value="paiement anticipét">paiement anticipét</option>
+                  <option value="aucun paiement">aucun paiement</option>
+               </select>
+            </div>
+            
+           
+          </div>
+        
+        </div>
+      </div>
+    </div>
+  </section>
+
+</form>
+
+
+
+@endsection
+
+@section('page-js')
+
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.0.0/jquery.min.js"></script>
+<link href="https://cdnjs.cloudflare.com/ajax/libs/select2/4.0.3/css/select2.min.css" rel="stylesheet" />
+<script src="{{ asset('js/plugins/select2.min.js') }}"></script>
+
+
 <script>
-   window.laravel ={!! json_encode([
-     'token' => csrf_token(),
-     'url'   => url('/'),
-     'date'   => date('Y-m-d'),
-   
-   
-   ]) !!}
+  window.laravel ={!! json_encode([
+    'token' => csrf_token(),
+    'url'   => url('/'),
+    'date'   => date('Y-m-d'),
+  
+  
+  ]) !!}
 </script>
 
 <script src="{{ asset('js/order_script.js') }}"></script>
+
+
 @endsection
